test(transport): address review nits on async memory spike
- Move inline tests to sibling async_memory_tests.rs per CLAUDE.md item 13
- Drop redundant file-level cfgs (mod declaration in transport/mod.rs already
  gates on test+async)
- Pin tokio::test(flavor = "current_thread") so the producer task can't race
  ahead of the consumer's park, preserving waker-path coverage
